requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
requests==2.27.1
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
